requirements.txt
altair @ file:///tmp/build/80754af9/altair_1599835197802/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs @ file:///C:/miniconda3/conda-bld/attrs_1762356943133/work
beautifulsoup4==4.14.2
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Bottleneck @ file:///C:/miniconda3/conda-bld/bottleneck_1761938212172/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1761745259803/work
bs4==0.0.2
cachetools @ file:///C:/Users/dev-admin/buildout/croot/cachetools_1759958658165/work
certifi @ file:///C:/Users/dev-admin/buildout/croot/certifi_1759956381452/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832793693/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744988796/work
click==8.0.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1758162070/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
entrypoints @ file:///C:/Users/dev-admin/buildout/croot/entrypoints_1759980387324/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastjsonschema==2.21.2
gitdb @ file:///C:/Users/dev-admin/buildout/croot/gitdb_1759961742190/work
GitPython @ file:///C:/Users/dev-admin/buildout/croot/gitpython_1759964586670/work
greenlet==3.2.4
h11==0.16.0
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1762350976/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///C:/Users/dev-admin/buildout/croot/jinja2_1761938590194/work
jsonschema @ file:///C:/miniconda3/conda-bld/jsonschema_1762459329751/work
jsonschema-specifications @ file:///C:/miniconda3/conda-bld/jsonschema-specifications_1762788613712/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
MarkupSafe @ file:///C:/Users/dev-admin/buildout/croot/markupsafe_1759955189290/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mkl-service==2.5.2
mkl_fft @ file:///C:/miniconda3/conda-bld/mkl_fft_1761592920453/work
mkl_random @ file:///C:/miniconda3/conda-bld/mkl_random_1761593081220/work
narwhals @ file:///C:/Users/dev-admin/buildout/croot/narwhals_1761941333126/work
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numexpr @ file:///C:/miniconda3/conda-bld/numexpr_1762165726164/work
numpy @ file:///C:/miniconda3/conda-bld/numpy_and_numpy_base_1762528516533/work/dist/numpy-2.3.4-cp314-cp314-win_amd64.whl#sha256=43bd4e704e50bb4093d40ab2f06efc0a326cf9254b914c3510471aa94cb86373
OleFileIO_PL==0.31
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1762332435977/work/dist/pandas-2.3.3-cp314-cp314-win_amd64.whl#sha256=ac6b8c30836d425b7de9199ad67d79b5d4d36663122fd5a70d81bfc894c0da42
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pillow @ file:///C:/Users/dev-admin/buildout/croot/pillow_1759959416858/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
plotly @ file:///opt/conda/envs/env/conda-bld/plotly_1738091316609/work/packages/python/plotly/dist/plotly-6.0.0.tar.gz#sha256=0b7834e43d05901cd1945fec77e668f6f07505ff4658f1a07632bdb7a0e8ec91
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf @ file:///D:/bld/protobuf_1760392525433/work/bazel-bin/python/dist/protobuf-6.31.1-cp314-abi3-win_amd64.whl#sha256=d66ffefda4529f9bec4d303dfcfab3f745a6e2f4cee8066e2eb5527252bdc268
psutil @ file:///D:/bld/bld/rattler-build_psutil_1762092943/work
psycopg2==2.9.11
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==21.0.0
pycparser @ file:///C:/Users/dev-admin/buildout/croot/pycparser_1759955361075/work
pydeck @ file:///C:/Users/dev-admin/buildout/croot/pydeck_1761997772384/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753029698/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
pytz @ file:///C:/Users/dev-admin/buildout/croot/pytz_1759958332927/work
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487145/work
PyYAML @ file:///C:/Users/dev-admin/buildout/croot/pyyaml_1759956226960/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387025/work
referencing @ file:///C:/miniconda3/conda-bld/referencing_1762536906066/work
requests @ file:///C:/miniconda3/conda-bld/requests_1762359708170/work
rpds-py @ file:///C:/miniconda3/conda-bld/rpds-py_1762366636411/work
selenium==4.38.0
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1761776807357/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
tornado @ file:///D:/bld/tornado_1762506711320/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
trio==0.32.0
trio-websocket==0.12.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3 @ file:///C:/Users/dev-admin/buildout/croot/urllib3_1761939442569/work
watchdog @ file:///C:/Users/dev-admin/buildout/croot/watchdog_1759968437007/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
webdriver-manager==4.0.2
websocket-client==1.9.0
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746275387/work
wsproto==1.2.0
xlrd==2.0.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
